refactor(docs): tab dx
