+++ apps/docs/content/components/tabs/radius.raw.jsx
@@ -1,0 +1,17 @@
+import {Tabs, Tab} from "@nextui-org/react";
+
+export default function App() {
+  const radiusList = ["full", "lg", "md", "sm", "none"];
+
+  return (
+    <div className="flex flex-wrap gap-4">
+      {radiusList.map((radius) => (
+        <Tabs key={radius} aria-label="Tabs radius" radius={radius}>
+          <Tab key="photos" title="Photos" />
+          <Tab key="music" title="Music" />
+          <Tab key="videos" title="Videos" />
+        </Tabs>
+      ))}
+    </div>
+  );
+}
