+++ apps/docs/content/components/tabs/sizes.ts
@@ -1,24 +1,4 @@
-const App = `import {Tabs, Tab} from "@nextui-org/react";
-
-export default function App() {
-  const sizes = [
-    "sm",
-    "md",
-    "lg",
-  ];
-
-  return (
-    <div className="flex flex-wrap gap-4">
-      {sizes.map((size) => (
-        <Tabs key={size} size={size} aria-label="Tabs sizes">
-          <Tab key="photos" title="Photos"/>
-          <Tab key="music" title="Music"/>
-          <Tab key="videos" title="Videos"/>
-        </Tabs>
-      ))}
-    </div>
-  );
-}`;
+import App from "./sizes.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
